Move tests scripts to `_testing` module
